+++ RELEASE_NOTES.md
@@ -30,7 +30,12 @@ Prolog adapter and integration files
 
 * FIXED: The integration scripts check for an outdated Logtalk user directory.
 
+Installers and installation scripts
+-----------------------------------
+
 * FIXED: The Windows installer check for an outdated Logtalk user directory.
+
+* FIXED: The embedding scripts check for an outdated Logtalk user directory.
 
 
 3.100.0 - May 14, 2026